# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sarangi DMS API
  version: 1.0.0-rc2
  description: Backend API for the Sarangi Dental Management System.
  contact:
    name: Sarangi DMS Support
    email: support@example.com
servers:
  - url: https://api.dev.example.com
    description: Dev
  - url: https://api.staging.example.com
    description: Staging
  - url: https://api.example.com
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      description: >
        Authenticates a user with email and password and returns a token pair.
      security: []
      operationId: authLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: >
        Exchanges a valid refresh token for a new token pair.
      security: []
      operationId: authRefresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /rx-presets:
    get:
      tags: [Prescriptions]
      summary: Search prescription presets
      description: >
        Typeahead search over prescription presets.
      operationId: searchPrescriptionPresets
      parameters:
        - in: query
          name: query
          required: false
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Prescription presets
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PrescriptionPreset'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/doctors:
    post:
      tags: [Admin]
      summary: Create doctor user
      description: >
        ADMIN-only. Creates a doctor user + profile.
      operationId: adminCreateDoctor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCreateDoctorRequest'
      responses:
        '201':
          description: Doctor created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDoctorListItem'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags: [Admin]
      summary: List doctors
      description: ADMIN-only. Returns all doctor profiles.
      operationId: adminListDoctors
      responses:
        '200':
          description: Doctors list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminDoctorListItem'
  /admin/doctors/{doctorId}:
    patch:
      tags: [Admin]
      summary: Update doctor profile
      description: >
        ADMIN-only. Partially updates doctor profile fields.
      operationId: adminUpdateDoctor
      parameters:
        - in: path
          name: doctorId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateDoctorRequest'
      responses:
        '200':
          description: Doctor updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDoctorListItem'
        '400':
          description: Invalid doctor id or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Doctor not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/rx-presets:
    post:
      tags: [Admin]
      summary: Create prescription preset
      description: ADMIN-only. Creates a new prescription preset.
      operationId: adminCreateRxPreset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCreateRxPresetRequest'
      responses:
        '201':
          description: Prescription preset created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrescriptionPreset'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/rx-presets/{id}:
    patch:
      tags: [Admin]
      summary: Update prescription preset
      description: ADMIN-only. Partially updates a prescription preset.
      operationId: adminUpdateRxPreset
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateRxPresetRequest'
      responses:
        '200':
          description: Prescription preset updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrescriptionPreset'
        '400':
          description: Invalid id or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Preset not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        message:
          type: string
          description: Human-readable error message.
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Optional field-level validation errors.
        traceId:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
    LoginResponse:
      type: object
      required: [userId, role, tokens]
      properties:
        userId:
          type: string
        role:
          type: string
          enum: [RECEPTION, DOCTOR, ADMIN]
        tokens:
          $ref: '#/components/schemas/TokenPair'
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 20
    RefreshResponse:
      type: object
      required: [tokens]
      properties:
        tokens:
          $ref: '#/components/schemas/TokenPair'
    TokenPair:
      type: object
      required: [accessToken, refreshToken, expiresInSec, refreshExpiresInSec]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresInSec:
          type: integer
        refreshExpiresInSec:
          type: integer
...
    RxLine:
      type: object
      required: [medicine, dose, frequency, duration]
      properties:
        medicine:
          type: string
        dose:
          type: string
        frequency:
          type: string
          enum: [QD, BID, TID, QID, HS, PRN]
        duration:
          type: integer
          minimum: 1
          maximum: 365
        sig:
          type: string
        timing:
          type: string
          enum: [BEFORE_MEAL, AFTER_MEAL, ANY]
        notes:
          type: string
...
    PrescriptionPreset:
      type: object
      required: [id, name, lines, createdByUserId, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        lines:
          type: array
          items:
            $ref: '#/components/schemas/RxLine'
        tags:
          type: array
          items:
            type: string
        createdByUserId:
          type: string
        createdAt:
          type: integer
    AdminCreateRxPresetRequest:
      type: object
      required: [name, lines]
      properties:
        name:
          type: string
        lines:
          type: array
          items:
            $ref: '#/components/schemas/RxLine'
        tags:
          type: array
          items:
            type: string
    AdminUpdateRxPresetRequest:
      type: object
      properties:
        name:
          type: string
        lines:
          type: array
          items:
            $ref: '#/components/schemas/RxLine'
        tags:
          type: array
          items:
            type: string
    AdminCreateDoctorRequest:
      type: object
      required: [email, displayName, password, fullName, registrationNumber, specialization]
      properties:
        email:
          type: string
          format: email
        displayName:
          type: string
        password:
          type: string
          minLength: 8
        fullName:
          type: string
        registrationNumber:
          type: string
        specialization:
          type: string
        contact:
          type: string
    AdminUpdateDoctorRequest:
      type: object
      properties:
        fullName:
          type: string
        registrationNumber:
          type: string
        specialization:
          type: string
        contact:
          type: string
        active:
          type: boolean
    AdminDoctorListItem:
      type: object
      required:
        - doctorId
        - fullName
        - registrationNumber
        - specialization
        - active
        - createdAt
        - updatedAt
        - email
        - displayName
      properties:
        doctorId:
          type: string
        fullName:
          type: string
        registrationNumber:
          type: string
        specialization:
          type: string
        contact:
          type: string
        active:
          type: boolean
        createdAt:
          type: integer
        updatedAt:
          type: integer
        email:
          type: string
          format: email
        displayName:
          type: string
...